Back-end/requirements.txt
Flask==1.1.2
psycopg2-binary < 3.0.0
flask_cors